NavigationUtils + Refactor
- create NavigationUtils
- return audiotype there to dev
- refactor audio classes out of their own directory, for later refactoring (likely into own Navigation directory)
