+++ libtelemetry/src/main/java/com/mapbox/services/android/telemetry/MapboxTelemetry.java
@@ -12,9 +12,6 @@ import android.net.NetworkInfo;
 import android.os.IBinder;
 import android.support.v4.content.LocalBroadcastManager;
 
-import com.mapbox.services.android.telemetry.audio.AudioTypeChain;
-import com.mapbox.services.android.telemetry.audio.AudioTypeResolver;
-
 import java.util.List;
 
 import okhttp3.Callback;
@@ -280,11 +277,4 @@ public class MapboxTelemetry implements FullQueueCallback, EventCallback {
       serviceBound = false;
     }
   };
-
-  public String obtainAudioType() {
-    AudioTypeChain audioTypeChain = new AudioTypeChain();
-    AudioTypeResolver setupChain = audioTypeChain.setup();
-
-    return setupChain.obtainAudioType(context);
-  }
 }
