+++ libtelemetry/src/main/java/com/mapbox/services/android/telemetry/NavigationUtils.java
@@ -1,0 +1,13 @@
+package com.mapbox.services.android.telemetry;
+
+import android.content.Context;
+
+public class NavigationUtils {
+
+  public static String obtainAudioType(Context context) {
+    AudioTypeChain audioTypeChain = new AudioTypeChain();
+    AudioTypeResolver setupChain = audioTypeChain.setup();
+
+    return setupChain.obtainAudioType(context);
+  }
+}
